+++ lib/src/interop/util/evented_interop.dart
@@ -9,17 +9,21 @@ typedef ListenerJsImpl = dynamic Function(EventJsImpl object);
 @JS()
 @anonymous
 abstract class EventJsImpl {
+  external String get id;
   external String get type;
   external LngLatJsImpl get lngLat;
   external List<FeatureJsImpl> get features;
   external PointJsImpl get point;
 
   external factory EventJsImpl({
+    String id,
     String type,
     LngLatJsImpl lngLat,
     List<FeatureJsImpl> features,
     PointJsImpl point,
   });
+
+  external preventDefault();
 }
 
 @JS('Evented')